fix(onboarding): rename organization-generator to camelCase convention #69
